refactor(learning circle):styles name  fixed
